+++ src/util.ts
@@ -2,3 +2,4 @@ export { figure } from './util/figure';
 export { footnote } from './util/footnote';
 export { toc } from './util/toc';
 export { info } from './util/info';
+export { context } from './util/context';